requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
google-generativeai
python-dotenv